+++ theory/src/main/java/at/ainf/theory/storage/AxiomSet.java
@@ -2,6 +2,7 @@ package at.ainf.theory.storage;
 
 import at.ainf.theory.watchedset.WatchedElement;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 /**
@@ -11,7 +12,7 @@ import java.util.Set;
  * Time: 09:31
  * To change this template use File | Settings | File Templates.
  */
-public interface AxiomSet<Id> extends Set<Id>, Comparable<AxiomSet<Id>>, WatchedElement<Double> {
+public interface AxiomSet<Id> extends Set<Id>, Comparable<AxiomSet<Id>>, WatchedElement<BigDecimal> {
 
     void updateAxioms(Set<Id> axioms);
 
@@ -25,9 +26,9 @@ public interface AxiomSet<Id> extends Set<Id>, Comparable<AxiomSet<Id>>, Watched
 
     void setValid(boolean valid);
 
-    void setMeasure(double value);
+    void setMeasure(BigDecimal value);
 
-    double getMeasure();
+    BigDecimal getMeasure();
 
     Set<Id> getEntailments();
